Move some code around and improve validation
