+++ custom_components/auth_oidc/tools/types.py
